@@ -46,7 +46,7 @@ class MultiColumnWizardExtension extends Extension
     /**
      * {@inheritdoc}
      */
-    public function getAlias()
+    public function getAlias(): string
     {
         return 'multicolumnwizard-bundle';
     }
