@@ -3,7 +3,7 @@
 /**
  * This file is part of menatwork/contao-multicolumnwizard-bundle.
  *
- * (c) 2012-2019 MEN AT WORK.
+ * (c) 2012-2023 MEN AT WORK.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,9 +13,10 @@
  * @package    menatwork/contao-multicolumnwizard-bundle
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
+ * @author     Ingolf Steinhardt <info@e-spin.de>
  * @copyright  2011 Andreas Schempp
  * @copyright  2011 certo web & design GmbH
- * @copyright  2013-2019 MEN AT WORK
+ * @copyright  2013-2023 MEN AT WORK
  * @license    https://github.com/menatwork/contao-multicolumnwizard-bundle/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -24,3 +25,4 @@ $GLOBALS['TL_LANG']['ERR']['mcwMinCount'] = 'Please fill in at least %s rows!';
 $GLOBALS['TL_LANG']['ERR']['mcwMaxCount'] = 'At most %s rows are allowed!';
 $GLOBALS['TL_LANG']['MSC']['tw_rnew']     = 'Ad a new row.';
 $GLOBALS['TL_LANG']['MSC']['tw_rmove']    = 'Move the row.';
+$GLOBALS['TL_LANG']['MSC']['fileManager'] = 'File manager';
